fix: move filterable attributes constant out of mocked module
