+++ src/services/item/plugins/publication/published/plugins/search/search.constants.ts
@@ -1,0 +1,12 @@
+import { TagCategory } from '@graasp/sdk';
+
+export const FILTERABLE_ATTRIBUTES = [
+  'isPublishedRoot',
+  'isHidden',
+  'lang',
+  'likes',
+  'creator',
+  TagCategory.Level,
+  TagCategory.Discipline,
+  TagCategory.ResourceType,
+] as const;
